Move test to where it can actually run
